+++ src/menu-items/index.js
@@ -1,14 +1,12 @@
 // project import
-import pages from './pages';
-import dashboard from './dashboard';
-import utilities from './utilities';
-import support from './support';
-import projectMenu from './projectMenu';
+import attendanceManagerMenu from './attendanceManagerMenu';
+import managerMenu from './managerMenu';
+import userMenu from './userMenu';
 
 // ==============================|| MENU ITEMS ||============================== //
 
 const menuItems = {
-  items: [projectMenu, dashboard, pages, utilities, support]
+  items: [userMenu, attendanceManagerMenu, managerMenu]
 };
 
 export default menuItems;
